Safari-609.1.11.1/LayoutTests/js/generational-opaque-roots-expected.txt

Tests that opaque roots behave correctly during young generation collections.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elem.customProperty is "hello"
PASS successfullyParsed is true

TEST COMPLETE
